Merged source of 3ds and PC Builds
There are still some bugs remaining, mainly with the textures. The textures are automatically resized in the 3ds build to not eat up memory, but this means that the texture sizes are wrong in one of the builds most of the time. I need to implement some kinda scaling variable right into the Texture2D class, so that the drawing functions can just use the scale of the objects and automatically correct the proper scaling. There is also the thing with the VSYNC and Frame Limiting, the Game Speed and such... Additionally, the mutexes are messed up, exiting the game while playing a map freezes the game, but unloading the game with Backspace and then exiting works every time??? IDK... Also, again with the textures, I need to create a proper system for managing the sizes of the skin objects, since osu skins suck, and come in different skin sizes..?? Also a rename/rebrand is a must copyrighted :P